requirements.txt
numpy==1.26.4
pandas==2.2.3
matplotlib==3.9.2
pyarrow==17.0.0
scipy ==1.14.1
dtw-python==1.5.3
fastdtw == 0.3.4
pytest == 8.3.4